\name{meSuSie-package}
\alias{meSuSie-package}
\alias{meSuSie}
\docType{package}
\title{\packageTitle{meSuSie}}
\description{\packageDescription{meSuSie}}
\section{Package Content}{\packageIndices{meSuSie}}
\author{\packageAuthor{meSuSie}}
\section{Maintainer}{\packageMaintainer{meSuSie}}
\keyword{package}
